requirements.txt
Pillow==2.6.1
beautifulsoup4==4.3.2
requests==2.5.1
